+++ packages/ui-components/src/__tests__/InputRegistryUrl.test.ts
@@ -4,22 +4,9 @@ import InputRegistryUrl from '../lib/components/input/InputRegistryUrl.svelte';
 import userEvent from '@testing-library/user-event';
 import { loadRegistryUrl } from '$lib/services/loadRegistryUrl';
 import type { RegistryManager } from '$lib/providers/registry/RegistryManager';
+import { initialRegistry } from '../__fixtures__/RegistryManager';
 
 const { mockRegistryStore } = await vi.hoisted(() => import('../lib/__mocks__/stores'));
-const mockDefaultRegistry = 'https://example.com/default-registry.json';
-let mockCurrentRegistry: string | null = mockDefaultRegistry; // Start with default
-
-export const initialRegistry: Partial<RegistryManager> = {
-	getCurrentRegistry: vi.fn(() => mockCurrentRegistry ?? mockDefaultRegistry),
-	setRegistry: vi.fn((newRegistry: string) => {
-		mockCurrentRegistry = newRegistry;
-	}),
-	resetToDefault: vi.fn(() => {
-		mockCurrentRegistry = mockDefaultRegistry;
-	}),
-	updateUrlWithRegistry: vi.fn(),
-	isCustomRegistry: vi.fn(() => mockCurrentRegistry !== mockDefaultRegistry)
-};
 
 vi.mock('../lib/services/loadRegistryUrl', () => ({
 	loadRegistryUrl: vi.fn()
